Fuel Counter by Wojciech Sobierajski
This is an application which counts fuel consumption, costs etc. of your car. 
It creates excel file to store data based on the inserted name. Input data is taken from Tkinter window and also displayed in Tkinter. 
Every time when opening the app and inserting the name, data base is being updated.